requirements.txt
py-cord[voice]>=2.6.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
faster-whisper>=1.0.0

# Optional — only needed for Spotify support
spotipy>=2.23.0
